+++ include/bahnhof/common/gamestate.h
@@ -1,5 +1,4 @@
 #pragma once
-#include<iostream>
 #include<SDL.h>
 #include<SDL_image.h>
 #include<SDL_ttf.h>
