+++ include/bahnhof/graphics/rendering.h
@@ -2,7 +2,6 @@
 #include<SDL.h>
 #include<SDL_image.h>
 #include<SDL_ttf.h>
-#include<string>
 #include "bahnhof/common/math.h"
 #include "bahnhof/common/orientation.h"
