+++ include/bahnhof/graphics/sprite.h
@@ -3,7 +3,6 @@
 #include<SDL_image.h>
 #include<SDL_ttf.h>
 #include<map>
-#include<string>
 #include "bahnhof/common/math.h"
 #include "bahnhof/common/orientation.h"
 #include "bahnhof/graphics/spritenames.h"
