+++ include/bahnhof/ui/buttons.h
@@ -2,7 +2,6 @@
 #include<SDL.h>
 #include<SDL_image.h>
 #include<SDL_ttf.h>
-#include<iostream>
 #include "bahnhof/common/math.h"
 #include "bahnhof/ui/ui.h"
