+++ include/bahnhof/ui/tables.h
@@ -2,7 +2,6 @@
 #include<SDL.h>
 #include<SDL_image.h>
 #include<SDL_ttf.h>
-#include<iostream>
 #include "bahnhof/ui/ui.h"
 #include "bahnhof/common/math.h"
 #include "bahnhof/rollingstock/trainmanager.h"
