+++ src/graphics/graphics.cpp
@@ -1,6 +1,4 @@
 #pragma once
-#include<iostream>
-#include<string>
 #include "bahnhof/graphics/graphics.h"
 #include "bahnhof/graphics/rendering.h"
