+++ src/rollingstock/rollingstock.cpp
@@ -1,5 +1,3 @@
-#include<iostream>
-#include<string>
 #include<map>
 #include "bahnhof/common/gamestate.h"
 #include "bahnhof/common/shape.h"
