+++ src/track/signal.cpp
@@ -1,5 +1,3 @@
-#include<iostream>
-#include<string>
 #include<map>
 #include "bahnhof/common/gamestate.h"
 #include "bahnhof/rollingstock/train.h"
