+++ src/track/trackinput.cpp
@@ -1,5 +1,3 @@
-#include<iostream>
-#include<string>
 #include<map>
 #include "bahnhof/routing/routing.h"
 #include "bahnhof/track/trackinternal.h"
