+++ src/ui/table.cpp
@@ -1,4 +1,3 @@
-#include<iostream>
 #include<format>
 #include "bahnhof/ui/ui.h"
 #include "bahnhof/ui/tables.h"
